requirements.txt
pymongo[srv]==4.2.0
from-root
pandas
scikit-learn
PyYAML
scipy
dill
imblearn
evidently
-e .